refactor: rename Checker → LoopChecker, inline types.Duration
Checker → LoopChecker:
- LoopChecker replaces Checker struct (collision with check.Checker)
- Updated tests (loop_test.go)

Inline types.Duration into internal:
- Moved from internal/types/ into internal/duration.go
- Deleted internal/types/ directory
- Updated conf.go, conf_validate.go, conf_test.go to use Duration directly
