+++ internal/duration_test.go
@@ -1,4 +1,4 @@
-package types
+package internal
 
 import (
 	"testing"
